backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
sqlmodel==0.0.24
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
email-validator==2.3.0
